+++ nextgen/console-reporter/src/default_reporter.rs
@@ -1,0 +1,1 @@
+pub struct DefaultReporter;
