+++ nextgen/console-reporter/src/lib.rs
@@ -1,0 +1,3 @@
+mod default_reporter;
+
+pub use default_reporter::*;
