+++ nextgen/console/src/lib.rs
@@ -1,6 +1,10 @@
+mod buffer;
 mod console;
 mod printer;
 pub mod prompts;
+mod reporter;
 
+pub use buffer::*;
 pub use console::*;
 pub use printer::*;
+pub use reporter::*;
